requirements.txt
fastapi[standard]==0.115.5
sqlalchemy==2.0.42
alembic==1.16.4
pymysql==1.1.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
uvicorn[standard]==0.32.1
pillow==10.4.0
aiofiles==24.1.0
pydantic==2.9.2
pydantic-settings==2.6.1
cryptography==43.0.3
mysql-connector-python==9.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.10.0
isort==5.13.2
mako==1.3.10
markupsafe==3.0.2
flake8==7.1.1
pytz==2025.2
